requirements.txt
# Core Dependencies
python-dotenv>=1.0.0
requests>=2.31.0
httpx>=0.25.0

# MCP Framework
mcp[cli]
fastmcp  # If mcp[cli] doesn't include this

# Graph Database
neo4j>=5.0.0

# Vector Database and Embeddings
chromadb>=0.4.22
sentence-transformers>=2.2.2

# PyTorch (let pip find compatible version for your system)
torch>=2.2.0

# Hugging Face Transformers (for NER and PICO classification)
transformers>=4.36.0

# Data Processing
pandas>=2.1.0
numpy>=1.26.0
biopython>=1.81

# API Utilities
xmltodict>=0.13.0

# Web Interface
streamlit>=1.28.0
streamlit-agraph>=0.0.40

# Graph Visualization
networkx>=3.2.1
matplotlib>=3.8.0
plotly>=5.17.0

# LLM Integration
ollama>=0.1.6

# Data Validation
pydantic>=2.0.0

# Logging and Utilities
loguru>=0.7.2
tqdm>=4.66.0

# Testing (optional, for test suite)
pytest>=7.4.0
pytest-cov>=4.1.0

# Note: We intentionally avoid TensorFlow/Keras to prevent M4 Mac conflicts
# The M4 GPU is utilized by Ollama for LLM inference
